requirements.txt
Click==8.0
Flask==2.1.2
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
google-generativeai==0.8.0
pymongo==4.10.1
flask_caching==2.0.0
google-cloud-storage

